Prototype for presentation PDF done.
Some things to note, it's a prototype. It's not fully working because of issues such as us not having a proper bucket to fetch blobs from nor having enough information from InfoOOH to put on the PDF. Also, I've downloaded React-PDF for some tests, but it has proven to be surprisingly difficult to work with, not to mention feels like it does the same thing, if not worse, as pdfmake. Will, at some point, probbaly continue messing with this, but for now, while I wait on others, I will solve the bucket problem, starting by researching Vercel.